+++ example.js
@@ -16,10 +16,10 @@ const ifpa = new Sigaa({
 (async () => fs.writeFileSync('./ifpa2.json', JSON.stringify(await ifpa.getCourses())))();
 
 // UFOPA
-const ufopa = new Sigaa({
-  institution: 'UFOPA',
-  debug: true,
-  verifyVersion: true,
-});
+// const ufopa = new Sigaa({
+//   institution: 'UFOPA',
+//   debug: true,
+//   verifyVersion: true,
+// });
 
-(async () => fs.writeFileSync('./ufopa.json', JSON.stringify(await ufopa.getCourses())))();
+// (async () => fs.writeFileSync('./ufopa.json', JSON.stringify(await ufopa.getCourses())))();