+++ CHANGELOG.md
@@ -2,14 +2,14 @@
 
 ### Added:
 
-- Changelog file ([a170972](https://github.com/viniciusaportela/get-sigaa/commit/))
+- Changelog file ([a170972](https://github.com/viniciusaportela/get-sigaa/commit/a170972d0ec91d66f1eed9cf0b0aca21a0949c06))
 - Now gives ID from course if it has describe in Name column ([0900cca](https://github.com/viniciusaportela/get-sigaa/commit/0900ccaac2086b3b1d2249d67b7155991f0393d4))
 - Check if package is up-to-date with sigaa website ([c1814a7](https://github.com/viniciusaportela/get-sigaa/commit/c1814a7a2cd1ea99cab429eb0b337512106b5571))
 - Added timer to debug inside Sigaa ([833ab22](https://github.com/viniciusaportela/get-sigaa/commit/833ab221a3bb60e95e7b296b7bb3a436f64bf63d))
 
 ### Fixed:
 
-- VSCode not hinting params to Sigaa construtor and utils/processTable ([a170972](https://github.com/viniciusaportela/get-sigaa/commit/))
+- VSCode not hinting params to Sigaa construtor and utils/processTable ([a170972](https://github.com/viniciusaportela/get-sigaa/commit/a170972d0ec91d66f1eed9cf0b0aca21a0949c06))
 
 ## 0.1.0 (12 February, 2020)
 ### Initial Version